+++ lib/controllers/quiz controllers/quiz_controller.dart
@@ -1,13 +1,12 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:fiservonboardingexp/util/constants.dart';
 import 'package:fiservonboardingexp/controllers/quiz%20controllers/auth_controller.dart';
-import 'package:fiservonboardingexp/screens/task%20pages/quiz%20screens/quiz_question_screen.dart';
 import 'package:fiservonboardingexp/widgets/quiz%20widgets/quiz_info_square.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:google_fonts/google_fonts.dart';
 
-import '../../util/elle_testing/firebase_ref/firebase_ref.dart';
+import '../../firebase references/firebase_refs.dart';
 import '../../model/quiz_model.dart';
 import '../../screens/task pages/quiz screens/question_screen.dart';
 import '../../firebase references/services/firebase_storage_service.dart';
