+++ lib/controllers/read_controller.dart
@@ -1,7 +1,8 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:fiservonboardingexp/util/kt_testing/firebase_reference.dart';
 import 'package:fiservonboardingexp/model/read_model.dart';
 import 'package:get/get.dart';
+
+import '../firebase references/firebase_refs.dart';
 
 class ReadController extends GetxController {
   late ReadModel document;
