+++ lib/routes/quiz_test_routes.dart
@@ -6,9 +6,9 @@ import 'package:fiservonboardingexp/screens/task%20pages/quiz%20screens/quiz_que
 import 'package:fiservonboardingexp/screens/task%20pages/quiz%20screens/splash/spash_screen.dart';
 import 'package:get/get.dart';
 
-import '../../../controllers/quiz controllers/question_controller.dart';
-import '../../../screens/task pages/quiz screens/introduction.dart';
-import '../../../screens/task pages/quiz screens/quiz_outcome_screen.dart';
+import '../controllers/quiz controllers/question_controller.dart';
+import '../screens/task pages/quiz screens/introduction.dart';
+import '../screens/task pages/quiz screens/quiz_outcome_screen.dart';
 
 class AppRoutes {
   static List<GetPage> routes() => [
