+++ lib/screens/menu drawer/change_password_page.dart
@@ -2,9 +2,9 @@ import 'package:fiservonboardingexp/themes/theme_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
-import '../firebase references/firebase_refs.dart';
-import '../widgets/app_bar_overlay.dart';
-import '../widgets/nav_bar.dart';
+import '../../firebase references/firebase_refs.dart';
+import '../../widgets/app_bar_overlay.dart';
+import '../../widgets/nav_bar.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 
 class ChangePasswordPage extends StatefulWidget {
