+++ lib/widgets/quiz widgets/quiz_app_bar.dart
@@ -1,5 +1,5 @@
 import 'package:fiservonboardingexp/util/constants.dart';
-import 'package:fiservonboardingexp/util/elle_testing/themes/ui_parameters.dart';
+import 'package:fiservonboardingexp/themes/quiz%20themes/ui_parameters.dart';
 import 'package:flutter/material.dart';
 
 class QuizAppBar extends StatelessWidget implements PreferredSizeWidget {
